修正公車伺服器掛掉時，不會閃退的問題

+++ NTOU_Mobile/module/ShuttleTrack/R66LeftLayerViewController.m
@@ -105,7 +105,7 @@
     return @"平常日發車時間";
 }
 
--(float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
+-(CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
 {
     return 32;
 }
